+++ esp-hal-procmacros/src/lib.rs
@@ -46,6 +46,7 @@
 
 use proc_macro::TokenStream;
 
+mod alert;
 mod blocking;
 mod builder;
 #[cfg(feature = "embassy")]
@@ -267,4 +268,42 @@ pub fn blocking_main(args: TokenStream, input: TokenStream) -> TokenStream {
 #[proc_macro_derive(BuilderLite, attributes(builder_lite))]
 pub fn builder_lite_derive(item: TokenStream) -> TokenStream {
     builder::builder_lite_derive(item)
+}
+
+/// Print a build error and terminate the process.
+///
+/// It should be noted that the error will be printed BEFORE the main function
+/// is called, and as such this should NOT be thought analogous to `println!` or
+/// similar utilities.
+///
+/// ## Example
+///
+/// ```rust
+/// esp_hal_procmacros::error! {"
+/// ERROR: something really bad has happened!
+/// "}
+/// // Process exits with exit code 1
+/// ```
+#[proc_macro]
+pub fn error(input: TokenStream) -> TokenStream {
+    alert::do_alert(termcolor::Color::Red, input);
+    panic!("Build failed");
+}
+
+/// Print a build warning.
+///
+/// It should be noted that the warning will be printed BEFORE the main function
+/// is called, and as such this should NOT be thought analogous to `println!` or
+/// similar utilities.
+///
+/// ## Example
+///
+/// ```rust
+/// esp_hal_procmacros::warning! {"
+/// WARNING: something unpleasant has happened!
+/// "};
+/// ```
+#[proc_macro]
+pub fn warning(input: TokenStream) -> TokenStream {
+    alert::do_alert(termcolor::Color::Yellow, input)
 }