+++ esp-storage/src/stub.rs
@@ -24,6 +24,7 @@ fn check<const ALIGN: u32, const SIZE: u32, const MAX_LEN: u32>(
     offset: u32,
     length: u32,
     data: *const u32,
+    is_write: bool,
 ) -> bool {
     if offset % ALIGN > 0 {
         print_error!("Not aligned offset: {offset}");
@@ -45,12 +46,12 @@ fn check<const ALIGN: u32, const SIZE: u32, const MAX_LEN: u32>(
         print_error!("Length out of range: {length} > {MAX_LEN}");
         return false;
     }
-    let addr = unsafe { (data as *const u8).offset_from(ptr::null()) } as u32;
+    let addr = data as u32;
     if addr % ALIGN > 0 {
         print_error!("Not aligned data: {addr:#0x}");
         return false;
     }
-    if unsafe { FLASH_LOCK } {
+    if is_write && unsafe { FLASH_LOCK } {
         print_error!("Flash locked");
         return false;
     }
@@ -58,7 +59,7 @@ fn check<const ALIGN: u32, const SIZE: u32, const MAX_LEN: u32>(
 }
 
 pub(crate) fn spiflash_read(src_addr: u32, data: *mut u32, len: u32) -> i32 {
-    if check::<WORD_SIZE, FLASH_SIZE, SECTOR_SIZE>(src_addr, len, data) {
+    if check::<WORD_SIZE, FLASH_SIZE, SECTOR_SIZE>(src_addr, len, data, false) {
         maybe_with_critical_section(|| {
             let src = unsafe { slice::from_raw_parts_mut(data as *mut u8, len as _) };
             unsafe { src.copy_from_slice(&FLASH_DATA[src_addr as usize..][..len as usize]) };
@@ -77,7 +78,7 @@ pub(crate) fn spiflash_unlock() -> i32 {
 }
 
 pub(crate) fn spiflash_erase_sector(sector_number: u32) -> i32 {
-    if check::<1, NUM_SECTORS, 1>(sector_number, 1, ptr::null()) {
+    if check::<1, NUM_SECTORS, 1>(sector_number, 1, ptr::null(), true) {
         maybe_with_critical_section(|| {
             let dst_addr = sector_number * SECTOR_SIZE;
             let len = SECTOR_SIZE;
@@ -90,7 +91,7 @@ pub(crate) fn spiflash_erase_sector(sector_number: u32) -> i32 {
 }
 
 pub(crate) fn spiflash_write(dest_addr: u32, data: *const u32, len: u32) -> i32 {
-    if check::<WORD_SIZE, FLASH_SIZE, SECTOR_SIZE>(dest_addr, len, data) {
+    if check::<WORD_SIZE, FLASH_SIZE, SECTOR_SIZE>(dest_addr, len, data, true) {
         maybe_with_critical_section(|| {
             let dst = unsafe { slice::from_raw_parts(data as *const u8, len as _) };
             for (d, s) in unsafe { &mut FLASH_DATA[dest_addr as usize..][..len as usize] }